notebook: move ui code into its own file
